Small changes in the Inventory class to be able to count how many of specific items are there in the inventory. Also now the goal is faded if the items are not collected, however the images that I used are just for the test, they need to be changed.
