+++ core/src/main/java/com/rim/IceField/GameMain.java
@@ -52,6 +52,7 @@ public class GameMain extends BasicGame {
         blizzardGUI = new BlizzardGUI();
         healthPanelGUI = new HealthPanelGUI(players, 20, 450);
 
+        p.getInventory().addItem(new Flare());
         map = new Map();
         mapgui = new MapGUI(map);
         mapgui.initialise();
@@ -91,6 +92,7 @@ public class GameMain extends BasicGame {
         blizzardGUI.update();
 
         healthPanelGUI.render();
+
